pyproject.toml
[project]
name = "deeptactics-arena"
version = "0.0.1"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
license-files = ["LICENSE"]
dependencies = [
    "chess>=1.11.2",
    "opencv-python>=4.13.0.92",
    "pettingzoo>=1.25.0",
    "pygame>=2.6.1",
    "python-dotenv>=1.2.1",
    "setuptools<70",
    "stable-baselines3>=2.7.1",
    "tensorboard>=2.20.0",
    "tensorflow>=2.20.0",
    "torch>=2.10.0",
    "wandb>=0.25.0",
]


[dependency-groups]
dev = [
    "pre-commit>=4.2.0",
    "pytest>=8.4.1",
    "pytest-cov>=6.2.1",
]

docs = [
    "mkdocs>=1.6.1",
    "mkdocs-api-autonav>=0.3.0",
    "mkdocs-autoapi>=0.4.1",
    "mkdocs-awesome-nav>=3.1.2",
    "mkdocs-material>=9.6.16",
    "mkdocstrings>=0.30.0",
    "mkdocstrings-python>=1.16.12",
]

[tool.commitizen]
name = "cz_conventional_commits"
tag_format = "$version"
version_scheme = "pep440"
version_provider = "uv"
update_changelog_on_bump = true
major_version_zero = true
